Revert "actualizacion proveedor y ventas"
This reverts commit 1aaeed143f591d3b41ca0ce919d64ee24b839e52.

+++ Papeleria/src/main/java/co/edu/unbosque/Papeleria/controllers/ProveedorController.java
@@ -60,7 +60,7 @@ public class ProveedorController {
     }
     
 	@GetMapping("/list_provider/{id}")
-    public ResponseEntity<Optional> listProviderId(@PathVariable int id) {
+    public ResponseEntity<Optional> listProviderId(@PathVariable Long id) {
         Optional<Proveedor> logs = provDao.searchProvider(id);
 
         if (logs == null) {
